move getuid/getpid to extensions.cpp and add TODO
